Moved populatePlaylist, populateTrackinfo, bindMousewheel to RS.PlayerWindow
